Code rewrite, partial commit. [WARN: Breaks stuff]
This commit is the start of converting the code into the new structure
supporting also systems. The changes are not complete (making this partial
commit since I got sick a bunch of times... long story) but is enough to
progress from. Since it is not complete, it breaks everything. Do not try to
run anything on this branch expecting anything but errors.

Also started documentation.