Added Rust library as proof of concept
Just to try it out as a proof that it can solve the memory/slowdown problem I
created a project
[meval-caller](https://github.com/johanbluecreek/meval-caller) which builds a
library `libmeval_caller.so` included in this commit (the code that builds to
exactly this binary can be found in commit
932925ac3963b93f47a8faabc889b3ae22208f66 of that project).

After this commit, this branch does not compute fitness anymore, it just
computes the expressions in a bunch of points, similarly to fitness, but
unrelated to the actual fitness. The point this shows is wonderful though: The
calculation is done, and there is no slow down of the execution, hence this
means that the issue is resolved in principle.

What should probably be done now is to merge the [meval-caller] project in
here, well parts of it at least, or the concept of it, so that I have one
place to build the library from, and so the project is self-contained.