Crazy good TL_solver for ODE1
Tried to add a feature to the solver to simulate wrapping events, which is
used by TL for expressions that achieve max length to make them truncate. Here
the same concept is not applicable, but to "simulate" such a thing I added a
second gene, and a header operator mostly consisting of the "z" operator. But
when I applied `muthead()` with the combo-method on the population, my last
5 or 6 runs have finished at 10 or less iterations.

This is very unlike wrapping for in TL, and may only be working well for ODE1,
but I want to save this in a commit.