A commit way too large
Updated readme to include a little status section, to be able to organise my
efforts a bit easier. Since one of the things not yet tested were SODEs, I
added one and tried it. This triggered a bug which, after some searching, was
found in `crossover()` and was a very silly bug. But before I found it, I
thought it was in `parse_expr()`, because it was not correct anyway (the
example in the documentation for `parse_expr()` was(/is) wrong for example).
But now, SODEs appear to be working, so time for a commit.