Global variables and examples
Started to implement a more sane handling of global variables. It is still
awkward and not fully implemented correctly for all functions yet, but this is
partially there now.

I also added a `generate_examples.jl` which should be used for generating the
examples for each function if any function is changed. This can also be run as
a consistency check to see if any function breaks as a result of an edit.