Moved and fixed TL_solver
TL solver now works a bit better, and is a function instead of hard-coded. So
you can pass you de-file to it and it will start solving.

Also used the unsafe-operators, because I think those give a more fair analogy
to the TL algorithm. But it still needs some tuning.