Experiment results and const globals
A first complete experiment has been run, for ODE1 -- everything as in this
commit -- and this is the result:

Sucesses: 23
Min time: 6.9141387939453125e-6
Max time: 663.6504371166229
Avg time: 134.24847841262817
Min iter: 1
Max iter: 70
Avg iter: 15.130434782608695

out of 30 and with 100 iterations as maximum. To be compared to [TL] that has
8, 1453, 653 for min, max, and avg, respectively, but with no failures. This
appears to be way earlier than [TL], so it appears that Polish notation may
generate random points better, but make the genetic operations worse than BNF.
But this it ok progress.

I tried to change all global variables to be 'const', however, this caused
very random segfaults. Even segfaults at "line 0 of no file", so this seems to
be a problem of the compiler. I will try and see if this is an issue already
filed with Julia.